Need advice
Looking for advice on how to move the my_share of the bill on line 68
to the method can_i_afford_the_bill on line 27.

More specifically, the problem is that the TipCalculator class variable
owner is saved as a string and it throws and error when trying to tack
a method onto that string in line 68.